Final version of GyabzSS on Forms
I am now going to take what I've done so far and try to apply this to
WPF to make it better. This is essentially the same as the last commit
but with the addition of the ability to screenshot a region. The region
is still a little WIP but works fine, though the application exits after
a region screenshot because of troubles i was having so i made it just
exit.